Support inherited attributes
The problem initially manifested when you would try to set an attribute
value on an instance of an inherited class and a NoMethodError was
thrown when the attribute couldn't be found when installing the
accessors. Now when looking up attributes we'll look at all the
attributes in the inheritance chain.